Split commands from extension into separate file
